+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 - Make MagikGrammar more consistent
 - Add check EmptyBlock
 - Dont check ##-comments in CommentedCodeCheck + handle SYNTAX\_ERRORS
+- Refactor getting templated checks to CheckList.getTemplatedChecks()
 
 
 0.3.2 (2019-10-29)